requirements.txt
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
flask>=3.0.0 